feat: add pandera and new annotations file
