+++ op_tcg/backend/models/bq_classes.py
@@ -1,12 +1,15 @@
 import json
+import pandera as pa
 from abc import ABC, abstractmethod
 from datetime import datetime
 
 from google.cloud import bigquery
 from pydantic import Field
 
+
 from op_tcg.backend.etl.load import bq_insert_rows, get_or_create_table, bq_upsert_row
 from op_tcg.backend.models.base import SQLTableBaseModel
+from op_tcg.backend.utils.annotations import create_pandera_schema_from_pydantic
 
 
 class BQTableBaseModel(SQLTableBaseModel, ABC):
@@ -16,6 +19,10 @@ class BQTableBaseModel(SQLTableBaseModel, ABC):
     _dataset_id: str
     create_timestamp: datetime = Field(default_factory=datetime.now, description="Creation timestamp when the insert in BQ happened")
 
+
+    @classmethod
+    def paSchema(cls):
+        return create_pandera_schema_from_pydantic(cls)
 
     def insert_to_bq(self, client: bigquery.Client | None = None):
         """Adds a new row to BQ"""
