+++ op_tcg/backend/models/matches.py
@@ -1,4 +1,5 @@
 import pandas as pd
+import pandera as pa
 from pydantic import BaseModel, Field
 from datetime import datetime, date
 from enum import IntEnum
