+++ pyproject.toml
@@ -20,6 +20,7 @@ google-cloud-storage = "^2.16.0"
 requests = "^2.31.0"
 beautifulsoup4 = "^4.12.3"
 st-theme = "^1.2.2"
+pandera = "^0.19.3"
 
 
 [tool.poetry.group.crawler.dependencies]
